feat: grouped api endpoints into files
